implemented new time-track update algorithm
*On the server:*
```Maintain an object containing the date time of first client in a room and a list of sockets in that room e.g.

{
   "001-874": {
       "date": "0001-01-01 01:00AM",
        "sockets": []
    }
}

Where "001" is the user-id and "874" is the patient id ...

When a time-track session starts, the socket is added to the "sockets" array
However, when a session ends, the socket is removed from the "sockets" array …. If during this process, the "sockets" array is empty, the object with the key ("001-874") is removed from the collection

A setInterval function is maintained on the server, which sends the elapsed interval between the "date" value and the current time to the clients.
```

*On the client:*
```The elapsed time in seconds is received every 10 seconds and the time in "hh:mm:ss" is calculated and displayed.
The time is also incremented every second so it always updates, but it's refreshed by the server at intervals.

This way, the server is the single source of truth and when all sessions truly end for a client, the time is saved in the DB```